refactor(useChat): 将 handler 分离到外部模块
将 15 个事件 handler + 2 张注册表从 useChat.ts 提取到独立模块：
- useChat.memory.ts — 5 个记忆事件 handler + memoryHandlers 注册表
- useChat.handlers.ts — 10 个主事件 handler + turnHandlers 注册表
- useChat.ts 缩减至 517 行，聚焦连接/通道/持久化/composable 职责
