+++ web/src/composables/useChat.ts
@@ -1,9 +1,11 @@
-import { reactive, computed, watch, nextTick, type Ref } from 'vue'
-import type { ClientMessage, ServerEvent, ChatTurn, ToolCall, ThinkingBlock, TurnEvent, ContextUsage, TokenEvent, AnswerEvent, ErrorEvent, DoneEvent, AskUserEvent, MemoryToolEvent, MemoryToolStartEvent, MemoryToolEndEvent, MemoryToolErrorEvent, MemoryStartEvent, MemoryDoneEvent } from '@/types'
+import { reactive, computed, watch, type Ref } from 'vue'
+import type { ClientMessage, ServerEvent, ChatTurn, ToolCall, ThinkingBlock, TurnEvent, ContextUsage, AskUserEvent, MemoryToolEvent } from '@/types'
 import { refreshSessions, switchSession } from '@/composables/useSession'
 import { buildFlatMessage, buildTimestamp, parseReferences } from '@/utils/references'
 import type { ParsedRef } from '@/utils/references'
 import { getToken } from '@/api'
+import { memoryHandlers, type MemoryEventType } from './useChat.memory'
+import { turnHandlers } from './useChat.handlers'
 /** 匹配旧格式尾缀（用于 localStorage 迁移） */
 const TIME_SUFFIX_RE = /（\d{4}-\d{2}-\d{2} \w{3} \d{2}:\d{2}）$/
 
@@ -94,7 +96,7 @@ const turnsCache = loadAllTurnsFromStorage()
 
 // ── 多会话通道 ─────────────────────────────────────────────
 
-interface SessionChannel {
+export interface SessionChannel {
   ws: WebSocket | null
   connected: boolean
   isStreaming: boolean
@@ -147,7 +149,7 @@ function getOrCreateChannel(sid: string): SessionChannel {
   return channels.get(sid)!
 }
 
-function persistTurns(sid: string) {
+export function persistTurns(sid: string) {
   const ch = channels.get(sid)
   if (!ch) {
     console.warn(`[useChat:persist] 跳过保存 sid="${sid}": 通道不存在`)
@@ -229,43 +231,6 @@ export function ensureConnected(sid: string) {
 
 // ── 事件路由 ──────────────────────────────────────────────
 
-/** read_memories 是纯读取操作，前端无需显示其执行状态。 */
-function skipReadMemories(payload: { tool_name: string }): boolean {
-  return payload.tool_name === 'read_memories'
-}
-
-/** 后台记忆 consumer 事件类型。 */
-type MemoryEventType = 'memory_start' | 'memory_tool_start' | 'memory_tool_end' | 'memory_tool_error' | 'memory_done'
-
-/** 后台记忆 consumer 事件处理器签名。 */
-type MemoryEventHandler = (ch: SessionChannel, sid: string, event: ServerEvent) => void
-
-/** 记忆事件处理器注册表。新增记忆事件类型只需在此注册，调用方守卫自动覆盖。 */
-const memoryHandlers = new Map<MemoryEventType, MemoryEventHandler>([
-  ['memory_start', handleMemoryStart],
-  ['memory_tool_start', handleMemoryToolStart],
-  ['memory_tool_end', handleMemoryToolEnd],
-  ['memory_tool_error', handleMemoryToolError],
-  ['memory_done', handleMemoryDone],
-])
-
-/** 事件路由处理器签名（turn 已由调用方守卫保证存在）。 */
-type TurnEventHandler = (ch: SessionChannel, sid: string, turn: ChatTurn, event: ServerEvent) => void
-
-/** 事件路由处理器注册表。 */
-const turnHandlers = new Map<string, TurnEventHandler>([
-  ['thinking_start', handleThinkingStart],
-  ['token', handleToken],
-  ['thinking_end', handleThinkingEnd],
-  ['tool_start', handleToolStart],
-  ['tool_end', handleToolEnd],
-  ['tool_error', handleToolError],
-  ['answer', handleAnswer],
-  ['done', handleDone],
-  ['error', handleError],
-  ['ask_user', handleAskUser],
-])
-
 function handleEventForChannel(sid: string, event: ServerEvent) {
   const ch = channels.get(sid)
   if (!ch) return
@@ -314,237 +279,6 @@ function handleEventForChannel(sid: string, event: ServerEvent) {
   const turnHandler = turnHandlers.get(event.type)
   if (turnHandler) {
     turnHandler(ch, sid, turn, event)
-  }
-}
-
-/** 后台记忆 consumer 开始处理本轮对话：压入「处理中」占位条目。 */
-function handleMemoryStart(ch: SessionChannel, sid: string, event: ServerEvent): void {
-  const me = event as MemoryStartEvent
-  console.log(`[ltm-fe] memory_start session=${sid} turn_id=${me.payload.turn_id}`)
-  const targetTurn = findTurnByBackendId(ch, me.payload.turn_id)
-  if (!targetTurn) { console.log(`[ltm-fe] NO turn found for ${me.payload.turn_id}`); return }
-  if (!targetTurn.memoryEvents) targetTurn.memoryEvents = []
-  targetTurn.memoryEvents.push({
-    kind: 'memory_tool', name: 'memory_processing', input: '', output: null, elapsed: null, status: 'running',
-  })
-}
-
-/** 后台记忆 consumer 开始调用 CRUD 工具：压入 running 状态工具事件。 */
-function handleMemoryToolStart(ch: SessionChannel, sid: string, event: ServerEvent): void {
-  const me = event as MemoryToolStartEvent
-  if (skipReadMemories(me.payload)) return
-  console.log(`[ltm-fe] memory_tool_start session=${sid} turn_id=${me.payload.turn_id} tool=${me.payload.tool_name}`)
-  const targetTurn = findTurnByBackendId(ch, me.payload.turn_id)
-  if (!targetTurn) { console.log(`[ltm-fe] NO turn found for ${me.payload.turn_id}`); return }
-  targetTurn.memoryEvents?.push({
-    kind: 'memory_tool', name: me.payload.tool_name, input: me.payload.input,
-    output: null, elapsed: null, status: 'running',
-  })
-}
-
-/** 后台记忆 consumer 的 CRUD 工具执行完毕：更新匹配事件为 done。 */
-function handleMemoryToolEnd(ch: SessionChannel, sid: string, event: ServerEvent): void {
-  const me = event as MemoryToolEndEvent
-  if (skipReadMemories(me.payload)) return
-  console.log(`[ltm-fe] memory_tool_end session=${sid} turn_id=${me.payload.turn_id} tool=${me.payload.tool_name}`)
-  const targetTurn = findTurnByBackendId(ch, me.payload.turn_id)
-  if (!targetTurn) { console.log(`[ltm-fe] NO turn`); return }
-  const mt = findRunningMemoryTool(targetTurn.memoryEvents ?? [], me.payload.tool_name)
-  if (mt) { mt.output = me.payload.output; mt.elapsed = me.payload.elapsed; mt.status = 'done' }
-  if (ch.turns.includes(targetTurn as ChatTurn)) persistTurns(sid)
-}
-
-/** 后台记忆 consumer 的 CRUD 工具出错：更新匹配事件为 error。 */
-function handleMemoryToolError(ch: SessionChannel, sid: string, event: ServerEvent): void {
-  const me = event as MemoryToolErrorEvent
-  if (skipReadMemories(me.payload)) return
-  const targetTurn = findTurnByBackendId(ch, me.payload.turn_id)
-  if (!targetTurn) return
-  const mt = findRunningMemoryTool(targetTurn.memoryEvents ?? [], me.payload.tool_name)
-  if (mt) mt.status = 'error'
-  if (ch.turns.includes(targetTurn as ChatTurn)) persistTurns(sid)
-}
-
-/** 后台记忆 consumer 处理完毕：移除「处理中」占位，无实际工具事件时渲染 memory_review。 */
-function handleMemoryDone(ch: SessionChannel, sid: string, event: ServerEvent): void {
-  const me = event as MemoryDoneEvent
-  console.log(`[ltm-fe] memory_done session=${sid} turn_id=${me.payload.turn_id}`)
-  const targetTurn = findTurnByBackendId(ch, me.payload.turn_id)
-  if (!targetTurn) { console.log(`[ltm-fe] NO turn`); return }
-  // 移除「处理中」占位条目
-  const realEvents = (targetTurn.memoryEvents ?? []).filter(e => e.name !== 'memory_processing')
-  targetTurn.memoryEvents = realEvents
-  if (realEvents.length === 0) {
-    targetTurn.memoryEvents = [{
-      kind: 'memory_tool', name: 'memory_review', input: '', output: '', elapsed: null, status: 'done',
-    }]
-    console.log(`[ltm-fe] added memory_review`)
-  }
-  if (ch.turns.includes(targetTurn as ChatTurn)) persistTurns(sid)
-}
-
-/** thinking_start：压入思考块。 */
-function handleThinkingStart(ch: SessionChannel, _sid: string, turn: ChatTurn, _event: ServerEvent): void {
-  turn.events.push({ kind: 'thinking', tokens: '', done: false, becameAnswer: false })
-}
-
-/** token：追加到最后一个思考块。 */
-function handleToken(ch: SessionChannel, _sid: string, turn: ChatTurn, event: ServerEvent): void {
-  const lastThink = findLastThinking(turn.events)
-  if (lastThink) {
-    lastThink.tokens += (event as TokenEvent).payload.token
-  }
-}
-
-/** thinking_end：标记最后一个思考块完成。 */
-function handleThinkingEnd(ch: SessionChannel, _sid: string, turn: ChatTurn, event: ServerEvent): void {
-  const lastThink = findLastThinking(turn.events)
-  if (lastThink) {
-    lastThink.done = true
-  }
-}
-
-/** tool_start：压入 running 状态工具调用。 */
-function handleToolStart(ch: SessionChannel, sid: string, turn: ChatTurn, event: ServerEvent): void {
-  const tsEvent = event as { type: 'tool_start'; payload: { call_id: string; tool_name: string; input: string } }
-  console.log('[useChat] tool_start:', tsEvent.payload.tool_name, { input: tsEvent.payload.input, call_id: tsEvent.payload.call_id, session: sid })
-  turn.events.push({
-    kind: 'tool',
-    name: tsEvent.payload.tool_name,
-    input: tsEvent.payload.input,
-    output: null,
-    elapsed: null,
-    status: 'running',
-    callId: tsEvent.payload.call_id,
-  })
-}
-
-/** tool_end：更新匹配工具调用为 done，提取 tool_data，处理 ask_user 状态恢复。 */
-function handleToolEnd(ch: SessionChannel, sid: string, turn: ChatTurn, event: ServerEvent): void {
-  const teEvent = event as { type: 'tool_end'; payload: { call_id: string; tool_name: string; output: string; elapsed: number; tool_data?: Record<string, unknown> } }
-  // 精确匹配：优先用 call_id，降级用 heuristic
-  const tc = findToolByCallId(turn.events, teEvent.payload.call_id)
-    ?? findBestMatchingTool(turn.events, teEvent.payload.tool_name)
-  if (tc) {
-    tc.output = teEvent.payload.output
-    tc.elapsed = teEvent.payload.elapsed
-    tc.status = 'done'
-    if (teEvent.payload.tool_data) {
-      tc.toolData = teEvent.payload.tool_data
-      if (teEvent.payload.tool_name === 'task_tracker' && teEvent.payload.tool_data) {
-        ch.taskTrackerData = teEvent.payload.tool_data as Record<string, unknown>
-      }
-    }
-    console.log(`[useChat] tool_end: "${teEvent.payload.tool_name}"`, {
-      output_len: (teEvent.payload.output || '').length,
-      output_preview: (teEvent.payload.output || '').slice(0, 100),
-      has_tool_data: !!teEvent.payload.tool_data,
-      elapsed: teEvent.payload.elapsed,
-      session: sid,
-    })
-  }
-  // ask_user 工具执行完毕 → 用户已回应，回到工作态
-  if (ch.isAwaitingUser && teEvent.payload.tool_name === ch._awaitingToolName) {
-    ch.isAwaitingUser = false
-    ch._awaitingToolName = null
-  }
-}
-
-/** tool_error：更新匹配工具调用为 error。 */
-function handleToolError(ch: SessionChannel, _sid: string, turn: ChatTurn, event: ServerEvent): void {
-  const teEvent = event as { type: 'tool_error'; payload: { call_id: string; tool_name: string; error: string } }
-  const tc = findToolByCallId(turn.events, teEvent.payload.call_id)
-    ?? findBestMatchingTool(turn.events, teEvent.payload.tool_name)
-  if (tc) {
-    tc.status = 'error'
-  }
-}
-
-/** answer：标记思考块为 becameAnswer，设置 finalAnswer。 */
-function handleAnswer(ch: SessionChannel, _sid: string, turn: ChatTurn, event: ServerEvent): void {
-  const lastThink = findLastThinking(turn.events)
-  if (lastThink) {
-    lastThink.becameAnswer = true
-  }
-  turn.finalAnswer = (event as AnswerEvent).payload.content
-}
-
-/** done：finalize 当前轮次，持久化，刷新会话列表。 */
-function handleDone(ch: SessionChannel, sid: string, turn: ChatTurn, event: ServerEvent): void {
-  const de = event as DoneEvent
-  ch.isAwaitingUser = false
-  ch._awaitingToolName = null
-  if (de.payload.context_usage) {
-    ch.contextUsage = de.payload.context_usage
-  }
-  // 存储后端 turn_id，用于关联后台记忆 consumer 的事件
-  if (de.payload.turn_id) {
-    turn.turnId = de.payload.turn_id
-    console.log(`[ltm-fe] turnId set on turn.id=${turn.id}: ${turn.turnId}`)
-  }
-  const lastThink = findLastThinking(turn.events)
-  const trackBecame = lastThink?.becameAnswer
-  console.log(`[useChat:done] 会话 ${sid}: becameAnswer=${trackBecame}, events=${turn.events.length}, finalAnswer=${turn.finalAnswer?.slice(0, 50) ?? 'null'}`)
-  if (trackBecame) {
-    const turnToFinalize = turn
-    void nextTick(() => {
-      setTimeout(() => {
-        console.log(`[useChat:done] becameAnswer 分支执行 persist (会话 ${sid})`)
-        ch.turns.push(turnToFinalize)
-        ch.currentTurn = null
-        ch.isStreaming = false
-        if (!ch.privateMode) { persistTurns(sid) }
-      }, 420)
-    })
-  } else {
-    console.log(`[useChat:done] 直接分支执行 persist (会话 ${sid})`)
-    ch.turns.push(turn)
-    ch.currentTurn = null
-    ch.isStreaming = false
-    if (!ch.privateMode) { persistTurns(sid) }
-  }
-  void refreshSessions()  // 轮次结束，刷新会话列表以更新 message_count
-  // 子 Agent 完成 → 自动切回主会话
-  if (ch.parentSessionId) {
-    setTimeout(() => switchSession(ch.parentSessionId!), 500)
-  }
-}
-
-/** error：清除流式状态，设置错误信息。 */
-function handleError(ch: SessionChannel, _sid: string, _turn: ChatTurn, event: ServerEvent): void {
-  ch.isAwaitingUser = false
-  ch._awaitingToolName = null
-  ch.error = (event as ErrorEvent).payload.message
-  ch.isStreaming = false
-}
-
-/** ask_user：设置交互等待状态，在 running 工具上挂载 interaction。 */
-function handleAskUser(ch: SessionChannel, sid: string, turn: ChatTurn, event: ServerEvent): void {
-  const ae = event as AskUserEvent
-  ch.isAwaitingUser = true
-  ch._awaitingToolName = ae.payload.tool_name
-  console.log('[useChat] received ask_user event:', {
-    tool_name: ae.payload.tool_name,
-    question: ae.payload.question?.slice(0, 50),
-    mode: ae.payload.mode,
-    interaction_id: ae.payload.interaction_id,
-    session: sid,
-  })
-  const runningTool = findFirstRunningToolForInteraction(turn.events, ae.payload.tool_name)
-  console.log('[useChat] findFirstRunningToolForInteraction result:', runningTool ? {
-    name: runningTool.name,
-    status: runningTool.status,
-    has_interaction: !!runningTool.interaction,
-  } : 'NOT FOUND')
-  if (runningTool) {
-    runningTool.interaction = {
-      question: ae.payload.question,
-      mode: ae.payload.mode,
-      options: ae.payload.options,
-      interactionId: ae.payload.interaction_id,
-      submitted: false,
-      code: ae.payload.code,
-    }
   }
 }
 
@@ -700,7 +434,7 @@ export function useChat(sessionId: Ref<string>) {
 
 
 
-function findLastThinking(events: TurnEvent[]): ThinkingBlock | undefined {
+export function findLastThinking(events: TurnEvent[]): ThinkingBlock | undefined {
   for (let i = events.length - 1; i >= 0; i--) {
     if (events[i].kind === 'thinking') {
       return events[i] as ThinkingBlock
@@ -712,7 +446,7 @@ function findLastThinking(events: TurnEvent[]): ThinkingBlock | undefined {
 /**
  * 通过 call_id 精确查找 ToolCall（当多个同名工具并行运行时用于精准匹配）。
  */
-function findToolByCallId(events: TurnEvent[], callId: string): ToolCall | undefined {
+export function findToolByCallId(events: TurnEvent[], callId: string): ToolCall | undefined {
   for (const e of events) {
     if (e.kind === 'tool' && e.callId === callId) {
       return e as ToolCall
@@ -726,7 +460,7 @@ function findToolByCallId(events: TurnEvent[], callId: string): ToolCall | undef
  * 从前往后找第一个 running 且尚未分配 interaction 的工具。
  * 当多个同名工具并行运行时，确保每个 ask_user 事件绑定到正确的实例。
  */
-function findFirstRunningToolForInteraction(events: TurnEvent[], toolName: string): ToolCall | undefined {
+export function findFirstRunningToolForInteraction(events: TurnEvent[], toolName: string): ToolCall | undefined {
   for (let i = 0; i < events.length; i++) {
     const e = events[i]
     if (e.kind === 'tool' && e.name === toolName && e.status === 'running' && !e.interaction) {
@@ -742,7 +476,7 @@ function findFirstRunningToolForInteraction(events: TurnEvent[], toolName: strin
  * 2. 其次找有 interaction 的工具
  * 3. 最后退化为旧行为：从后往前找最后一个 running 的工具
  */
-function findBestMatchingTool(events: TurnEvent[], toolName: string): ToolCall | undefined {
+export function findBestMatchingTool(events: TurnEvent[], toolName: string): ToolCall | undefined {
   // 第一遍：优先匹配用户已提交响应的工具
   for (let i = 0; i < events.length; i++) {
     const e = events[i]
@@ -768,13 +502,13 @@ function findBestMatchingTool(events: TurnEvent[], toolName: string): ToolCall |
 }
 
 /** 通过后端 turn_id 查找 turn（先在 currentTurn 中找，再在 turns 中找）。 */
-function findTurnByBackendId(ch: SessionChannel, turnId: string): ChatTurn | undefined {
+export function findTurnByBackendId(ch: SessionChannel, turnId: string): ChatTurn | undefined {
   if (ch.currentTurn?.turnId === turnId) return ch.currentTurn
   return ch.turns.find(t => t.turnId === turnId)
 }
 
 /** 在 memoryEvents 中查找指定工具名的 running 状态事件。 */
-function findRunningMemoryTool(events: MemoryToolEvent[], toolName: string): MemoryToolEvent | undefined {
+export function findRunningMemoryTool(events: MemoryToolEvent[], toolName: string): MemoryToolEvent | undefined {
   for (let i = events.length - 1; i >= 0; i--) {
     const e = events[i]
     if (e.name === toolName && e.status === 'running') return e
